camp_spotter_backend/requirements.txt
cryptography==36.0.1
defusedxml==0.7.1
cffi==1.15.0
Pillow==9.0.1
pip==22.0.3
wheel==0.36.2
pytz==2021.3
Django==4.0.1
sqlparse==0.4.2
asgiref==3.5.0
certifi==2021.10.8
six==1.16.0
oauthlib==3.2.0
idna==3.3
urllib3==1.26.8
pycparser==2.21
setuptools==57.0.0
djangorestframework==3.13.1
requests==2.27.1